refactor: rename _mcmc_distribution.py to _approximate_distribution.py
The file holds ApproximateDistribution; the old name predates the
class rename and no longer matches its contents. Updates all
imports and the file-tree references in CONTRIBUTING.md and
STYLE_GUIDE.md.
