+++ probpipe/inference/_approximate_distribution.py
