+++ probpipe/inference/_cmdstan_method.py
@@ -8,7 +8,7 @@ import arviz as az
 import jax.numpy as jnp
 
 from ..core._registry import MethodInfo
-from ._mcmc_distribution import ApproximateDistribution, make_posterior
+from ._approximate_distribution import ApproximateDistribution, make_posterior
 from ._registry import InferenceMethod
